Rename org-project to org-projprop.
To fix the naming conflict with the other org-project, rename this to
org-projprop. This name was chosen to be relatively short and underscore this is
all about setting and using a project property.

IMPORTANT!
This increases the version to 1.0, because it is a breaking change.

To fix your config, please rename all instances of org-project to org-projprop.